Move type to type file + rebase
