+++ client/src/types/index.ts
@@ -4,3 +4,10 @@ export interface Health {
   current: bigint;
   lifetime: bigint;
 }
+
+export enum ResourceMiningTypes {
+  Forge = "forge",
+  Mine = "mine",
+  LumberMill = "lumber_mill",
+  Dragonhide = "dragonhide",
+}
