+++ client/src/ui/utils/utils.tsx
@@ -14,10 +14,11 @@ import {
   default as realmHexPositions,
   default as realmsHexPositions,
 } from "../../data/geodata/hex/realmHexPositions.json";
-import { FELT_CENTER, ResourceMiningTypes } from "../config";
+import { FELT_CENTER } from "../config";
 import { SortInterface } from "../elements/SortButton";
 import * as THREE from "three";
 import { HEX_HORIZONTAL_SPACING, HEX_VERTICAL_SPACING } from "@/three/scenes/HexagonScene";
+import { ResourceMiningTypes } from "@/types";
 
 export { getEntityIdFromKeys };
